requirements.txt
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.4
colorama==0.4.3
dj-database-url==0.5.0
Django==3.0.6
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-rest-multiple-models==2.1.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.0.1
gunicorn==20.0.4
isort==4.3.21
jdcal==1.4.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
openpyxl==3.0.4
pycodestyle==2.6.0
PyJWT==2.4.0
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
whitenoise==5.1.0
wrapt==1.12.1
XlsxWriter==1.3.5
